ステータスの描画責務をGameStatusに移す
