+++ src/model/tictactoe-interface.ts
@@ -1,3 +1,5 @@
 export type Player = "X" | "O"
 
 export type Marker = Player | "BLANK"
+
+export type Status = `WIN_X` | `WIN_O` | "DRAW" | `NEXT_X` | `NEXT_O`
